@@ -5,7 +5,9 @@ import androidx.lifecycle.viewModelScope
 import com.abbaraees.weatherly.R
 import com.abbaraees.weatherly.data.LocationResponse
 import com.abbaraees.weatherly.data.WeatherDataNetwork
-import com.abbaraees.weatherly.states.HomeState
+import com.abbaraees.weatherly.data.entities.WeatherData
+import com.abbaraees.weatherly.data.repository.WeatherDataRepository
+import com.abbaraees.weatherly.ui.states.HomeState
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.request.get
@@ -15,18 +17,27 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import java.time.Instant
 import java.time.ZoneId
-import java.util.Date
+import java.time.format.DateTimeFormatter
 
 
 class HomeViewModel(
-    private val httpClient: HttpClient
+    private val httpClient: HttpClient,
+    private val repository: WeatherDataRepository,
 ): ViewModel() {
 
     private val _state = MutableStateFlow(HomeState())
 
     val state = _state
 
-    fun fetchWeatherData(lat: Double, lon: Double) {
+    /**
+     * Fetches weather data based on the provided latitude and longitude.
+     * @param lat Latitude of the location.
+     * @param lon Longitude of the location.
+     * @param name Name of the location.
+     * @param state State of the location.
+     * @param country Country of the location.
+     */
+    fun fetchWeatherData(lat: Double, lon: Double, name: String, state: String, country: String) {
         _state.update { it.copy(
             isFetchingWeatherData = true,
             isLoadingLocations = false,
@@ -51,14 +62,16 @@ class HomeViewModel(
             if (response != null) {
                 if (response.status.value == 200) {
                     val data: WeatherDataNetwork = response.body()
+                    val formattedDate = Instant
+                        .ofEpochSecond(data.dt)
+                        .atZone(ZoneId.systemDefault())
+                        .toLocalDateTime().format(
+                        DateTimeFormatter.ofPattern("dd/MM/yyyy hh:mm a"))
                     _state.update {
                         it.copy(
                             cityName = data.name,
                             weatherDescription = data.weather[0].description,
-                            dateTime = Instant
-                                .ofEpochSecond(data.dt)
-                                .atZone(ZoneId.systemDefault())
-                                .toLocalDateTime().toString(),
+                            dateTime = formattedDate,
                             temp = data.main.temp,
                             isFetchingWeatherData = false,
                             weatherIcon = when (data.weather[0].main) {
@@ -70,6 +83,23 @@ class HomeViewModel(
                             }
                         )
                     }
+
+                    val weatherData = WeatherData(
+                        latitude = lat,
+                        longitude = lon,
+                        locationName = name,
+                        state = state,
+                        country = country,
+                        weatherMain = data.weather[0].main,
+                        weatherDescription = data.weather[0].description,
+                        temperature = data.main.temp,
+                        date = formattedDate
+                    )
+
+                    viewModelScope.launch {
+                        repository.addWeatherData(weatherData)
+                    }
+
                 }
             }
         }
