@@ -2,15 +2,19 @@ package com.abbaraees.weatherly.viewmodels
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
+import com.abbaraees.weatherly.data.repository.WeatherDataRepository
 import io.ktor.client.HttpClient
 
 class ViewModelFactory(
-    private val httpClient: HttpClient
+    private val httpClient: HttpClient,
+    private val repository: WeatherDataRepository,
 ): ViewModelProvider.NewInstanceFactory() {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(HomeViewModel::class.java))
-            return HomeViewModel(httpClient) as T
-        else
-            return super.create(modelClass)
+        return when {
+            modelClass.isAssignableFrom(HomeViewModel::class.java) -> {
+                HomeViewModel(httpClient, repository) as T
+            }
+            else -> OthersViewModel(repository) as T
+        }
     }
 }
